+++ tests/tests.rs
@@ -2,6 +2,7 @@
 extern crate log;
 
 use anyhow::Result;
+use bytes::Bytes;
 use futures::{
     future::FutureExt,
     stream::{futures_unordered::FuturesUnordered, FuturesOrdered, StreamExt, TryStreamExt},
@@ -79,8 +80,7 @@ async fn test_storage_read_write() {
     let data = b"test data string";
     write_one(&storage, 1234, data, None).await.unwrap();
     let new_data = storage.read(KeyTest::new(key)).await.unwrap();
-    assert!(matches!(new_data, ReadResult::Found(_)));
-    assert_eq!(new_data.unwrap(), data);
+    assert_eq!(new_data, ReadResult::Found(Bytes::copy_from_slice(data)));
     common::clean(storage, path).await.unwrap();
     warn!("elapsed: {:.3}", now.elapsed().as_secs_f64());
 }
@@ -325,7 +325,7 @@ async fn test_index_from_empty_blob() {
     assert!(blob_file_path.exists());
     let new_storage = common::create_test_storage(&path, 1_000_000).await.unwrap();
     new_storage
-        .write(KeyTest::new(1), vec![1; 8])
+        .write(KeyTest::new(1), vec![1; 8].into())
         .await
         .unwrap();
     new_storage.close().await.unwrap();
@@ -389,7 +389,10 @@ async fn test_write_512_records_with_same_key() {
         let mut meta = Meta::new();
         meta.insert("version".to_owned(), i.to_string());
         sleep(Duration::from_micros(1)).await;
-        storage.write_with(&key, value.clone(), meta).await.unwrap();
+        storage
+            .write_with(&key, value.clone().into(), meta)
+            .await
+            .unwrap();
     }
     common::clean(storage, path).await.expect("clean failed");
     warn!("elapsed: {:.3}", now.elapsed().as_secs_f64());
@@ -419,9 +422,12 @@ async fn test_read_with() {
     // data_read - last record, data_read_with - first record with "1.0" meta
     assert!(matches!(data_read_with, ReadResult::Found(_)));
     assert!(matches!(data_read, ReadResult::Found(_)));
-    let data_read_with = data_read_with.unwrap();
-    assert_ne!(data_read_with, data_read.unwrap());
-    assert_eq!(data_read_with, data1);
+    let data_read_with = data_read_with;
+    assert_ne!(data_read_with, data_read);
+    assert_eq!(
+        data_read_with,
+        ReadResult::Found(Bytes::copy_from_slice(data1))
+    );
     common::clean(storage, path).await.expect("clean failed");
     warn!("elapsed: {:.3}", now.elapsed().as_secs_f64());
 }
@@ -535,10 +541,10 @@ async fn test_check_bloom_filter_single() {
         let neg_key = KeyTest::new(i + 2 * repeat);
         trace!("key: {}, pos: {:?}, negative: {:?}", i, pos_key, neg_key);
         let key = KeyTest::new(i);
-        storage.write(&key, data.to_vec()).await.unwrap();
+        storage.write(&key, data.to_vec().into()).await.unwrap();
         assert_eq!(storage.check_filters(key).await, Some(true));
         let data = b"other_random_data";
-        storage.write(&pos_key, data.to_vec()).await.unwrap();
+        storage.write(&pos_key, data.to_vec().into()).await.unwrap();
         assert_eq!(storage.check_filters(pos_key).await, Some(true));
         assert_eq!(storage.check_filters(neg_key).await, Some(false));
     }
@@ -554,7 +560,7 @@ async fn test_check_bloom_filter_multiple() {
         b"lfolakfsjher_rladncreladlladkfsje_pkdieldpgkeolladkfsjeslladkfsj_slladkfsjorladgedom_dladlladkfsjlad";
     for i in 1..800 {
         let key = KeyTest::new(i);
-        storage.write(&key, data.to_vec()).await.unwrap();
+        storage.write(&key, data.to_vec().into()).await.unwrap();
         sleep(Duration::from_millis(6)).await;
         trace!("blobs count: {}", storage.blobs_count().await);
     }
@@ -577,7 +583,7 @@ async fn test_check_bloom_filter_multiple_offloaded() {
         b"lfolakfsjher_rladncreladlladkfsje_pkdieldpgkeolladkfsjeslladkfsj_slladkfsjorladgedom_dladlladkfsjlad";
     for i in 1..800 {
         let key = KeyTest::new(i);
-        storage.write(&key, data.to_vec()).await.unwrap();
+        storage.write(&key, data.to_vec().into()).await.unwrap();
         sleep(Duration::from_millis(6)).await;
         trace!("blobs count: {}", storage.blobs_count().await);
     }
@@ -606,7 +612,7 @@ async fn test_check_bloom_filter_init_from_existing() {
         for i in 1..base {
             let key = KeyTest::new(i);
             trace!("write key: {}", i);
-            storage.write(&key, data.to_vec()).await.unwrap();
+            storage.write(&key, data.to_vec().into()).await.unwrap();
             trace!("blobs count: {}", storage.blobs_count().await);
         }
         debug!("close storage");
@@ -652,7 +658,7 @@ async fn test_check_bloom_filter_generated() {
         for i in 1..base {
             let key = KeyTest::new(i);
             trace!("write key: {}", i);
-            storage.write(&key, data.to_vec()).await.unwrap();
+            storage.write(&key, data.to_vec().into()).await.unwrap();
             trace!("blobs count: {}", storage.blobs_count().await);
         }
         debug!("close storage");
@@ -692,7 +698,7 @@ async fn write_one(
     data: &[u8],
     version: Option<&str>,
 ) -> Result<()> {
-    let data = data.to_vec();
+    let data = Bytes::copy_from_slice(data);
     let key = KeyTest::new(key);
     debug!("tests write one key: {:?}", key);
     if let Some(v) = version {
@@ -1055,7 +1061,7 @@ async fn test_in_memory_and_disk_records_retrieval() -> Result<()> {
                     .into_iter()
                     .map(|e| e.load())
                     .collect::<FuturesOrdered<_>>()
-                    .map(|e| e.map(|r| r.into_data()))
+                    .map(|e| e.map(|r| r.into_data().into()))
                     .try_collect::<Vec<_>>()
             })
             .await;